clean up logic and partially parallelize build_knn_graph
Removed the usage of 'inf' values in the function, and further cleaned up the logic, to make it read in a more
deterministic manner. Extracted the distance calculating part of the function as another global function, so that it may be parallelized through the usage of numba.prange and numba.njit(parallel=True) flags.